Add the circle background layout and update the original Landing.tsx file.
